Small config refactor
