+++ packages/sourcecred/src/core/clearcred/config.js
@@ -2,7 +2,7 @@
 
 import type {TimestampMs} from "../../util/timestamp";
 import type {Contribution, Expression, WeightOperand} from "./contribution";
-import type {Operator} from "./operator";
+import {type Operator, OPERATOR_KEY_PREFIX, OPERATORS} from "./operator";
 
 /**
 Semantically, allows weight configuration of different qualities/characteristics of
@@ -43,7 +43,7 @@ operators within an expression. For example, one might be able to
 configure that emoji reactions be added or multiplied.
  */
 export type OperatorConfig = $ReadOnlyArray<{|
-  /** Arbitrary string enumerated by a plugin developer. */
+  /** Arbitrary string enumerated by a plugin developer. NOT prefixed. */
   +key: string,
   +operator: Operator,
 |}>;
@@ -74,44 +74,32 @@ export type Config = {|
 /**
 Returns true if the subkey exists in the subkeys array of the key.
 Returns false if the subkey does not exist in the subkeys array.
-Throws if the key is not found.
+Throws if the key has not been set in the configuration.
  */
 export function hasExplicitWeight(
   {key, subkey}: WeightOperand,
   config: Config
 ): boolean {
-  const keyConfig = config.weights.find(
-    (x) => x.key === key
-  );
+  const keyConfig = config.weights.find((x) => x.key === key);
   if (keyConfig === undefined)
-    throw new Error(
-      `Key [${key}] has not been set in the weights config.`
-    );
-  const subkeyConfig = keyConfig.subkeys.find(
-    (x) => x.subkey === subkey
-  );
+    throw new Error(`Key [${key}] has not been set in the weights config.`);
+  const subkeyConfig = keyConfig.subkeys.find((x) => x.subkey === subkey);
   return subkeyConfig !== undefined;
 }
 
 /**
 If the subkey is found, returns the subkey's weight.
 If the subkey is not found, returns the key's default.
-Throws if the key is not found.
+Throws if the key has not been set in the configuration.
  */
 export function getWeight(
   {key, subkey}: WeightOperand,
   config: Config
 ): number {
-  const keyConfig = config.weights.find(
-    (x) => x.key === key
-  );
+  const keyConfig = config.weights.find((x) => x.key === key);
   if (keyConfig === undefined)
-    throw new Error(
-      `Key [${key}] has not been set in the weights config.`
-    );
-  const subkeyConfig = keyConfig.subkeys.find(
-    (x) => x.subkey === subkey
-  );
+    throw new Error(`Key [${key}] has not been set in the weights config.`);
+  const subkeyConfig = keyConfig.subkeys.find((x) => x.subkey === subkey);
   if (!subkeyConfig) {
     return keyConfig.default;
   }
@@ -119,15 +107,42 @@ export function getWeight(
 }
 
 /**
-Returns the amount set for the key, or 0 if the key is not found.
+Takes a prefixed key and returns the configured operator queried by the
+non-prefixed key.
+Throws if the input is not properly prefixed.
+Throws if the key has not been set in the configuration.
+Throws if the configured operator is not a valid operator.
  */
-export function getShares(
-  key: string,
-  config: Config
-): number {
-  return (
-    config.shares.find(
-      (shareConfig) => shareConfig.key === key
-    )?.amount ?? 0
-  );
+export function getOperator(rawKey: string, config: Config): Operator {
+  if (!rawKey.startsWith(OPERATOR_KEY_PREFIX))
+    throw new Error(
+      `Invalid expression operator [${rawKey}]. This is probably a bug in the plugin. Valid operators are ${OPERATORS.toString()} and operator keys should be prefixed with '${OPERATOR_KEY_PREFIX}'`
+    );
+  const key = rawKey.slice(OPERATOR_KEY_PREFIX.length);
+  const operator = config.operators.find(
+    (operatorConfig) => key === operatorConfig.key
+  )?.operator;
+  if (!operator)
+    throw new Error(
+      `Operator for key [${key}] has not been set in the operators config.`
+    );
+  if (!OPERATORS.includes(operator))
+    throw new Error(
+      `Operator [${operator}] for Key [${key}] is an invalid configuration. Please choose from ${OPERATORS.toString()}.`
+    );
+  return operator;
+}
+
+/**
+Returns the amount set for the key.
+Throws if the key has not been set in the configuration.
+ */
+export function getShares(key: string, config: Config): number {
+  const amount = config.shares.find((shareConfig) => shareConfig.key === key)
+    ?.amount;
+  if (!amount)
+    throw new Error(
+      `Shares for key [${key}] has not been set in the operators config.`
+    );
+  return amount;
 }
